+++ src/UnidadMonetaria/UnidadMonetaria.cpp
@@ -3,13 +3,13 @@
 //
 
 #include "UnidadMonetaria.h"
-const std::pair<int, std::string> &UnidadMonetaria::getUnidad() const {
+const std::pair<double, std::string> &UnidadMonetaria::getUnidad() const {
   return Unidad;
 }
-void UnidadMonetaria::setUnidad(const std::pair<int, std::string> &unidad) {
+void UnidadMonetaria::setUnidad(const std::pair<double, std::string> &unidad) {
   Unidad = unidad;
 }
-UnidadMonetaria::UnidadMonetaria(int valor, const char simbolo[]) {
-  std::pair<int,std::string> unidad(valor,simbolo);
+UnidadMonetaria::UnidadMonetaria(double valor, const char simbolo[]) {
+  std::pair<double,std::string> unidad(valor,simbolo);
   this->Unidad = unidad;
 }
